Moved folders around to match standard go project structure
Created cmd and pkg as top level folder and moved folders around appropriately.  Fixed
up the makefiles and dockerfiles.
