--- cloud/vsphere/cmd/vsphere-machine-controller/.gitignore
@@ -1,1 +1,0 @@
-vsphere-machine-controller
