+++ cmd/clusterctl/examples/vsphere/.gitignore
