+++ cmd/clusterctl/examples/vsphere/addons.yaml.template
