+++ cmd/clusterctl/examples/vsphere/cluster.yaml.template
