+++ cmd/clusterctl/examples/vsphere/generate-yaml.sh
