+++ cmd/clusterctl/examples/vsphere/machines.yaml.template
