+++ cmd/clusterctl/examples/vsphere/machineset.yaml.template
