+++ cmd/clusterctl/examples/vsphere/provider-components.yaml.template
