+++ pkg/cloud/vsphere/constants/constants.go
