+++ pkg/cloud/vsphere/instancestatus.go
