+++ pkg/cloud/vsphere/namedmachines/namedmachines.go
