+++ pkg/cloud/vsphere/provisioner/common/templates.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"text/template"
 
-	vsphereutils "sigs.k8s.io/cluster-api-provider-vsphere/cloud/vsphere/utils"
+	vsphereutils "sigs.k8s.io/cluster-api-provider-vsphere/pkg/cloud/vsphere/utils"
 	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
 )
