+++ pkg/cloud/vsphere/provisioner/govmomi/provisioner.go
