+++ pkg/cloud/vsphere/vsphereproviderconfig/types.go
