+++ pkg/cloud/vsphere/vsphereproviderconfig/v1alpha1/types.go
