+++ pkg/cloud/vsphere/vsphereproviderconfig/v1alpha1/zz_generated.deepcopy.go
