+++ pkg/cloud/vsphere/vsphereproviderconfig/zz_generated.deepcopy.go
